+++ kperm/__init__.py
@@ -1,1 +1,1 @@
-from kperm.channel import Channel, detect_sf, detectSF
+from kperm.channel import Channel, detect_sf